75 percent done

+++ projects/simple-rest-api/controllers/todoController.js
@@ -19,17 +19,23 @@ const handleErrors = (error) => {
 };
 
 const todos_get = async (req, res) => {
+  const user = req.user;
   try {
-    const todos = await Todo.find().sort({ createdAt: -1 });
-    res.status(200).json({todos})
+    const todos = await Todo.find({ user }).sort({ createdAt: -1 });
+    if(!todos) {
+      res.status(200).json({todos: "Your todos are empty"})
+    }
+    else {
+      res.status(200).json({ todos });
+    }
   } catch (error) {
-    res.status(400).send("Error fetching data");
+    res.status(500).send("Error fetching data");
     console.log(error);
   }
 };
 
 const todos_post = async (req, res) => {
-  const user = req.user
+  const user = req.user;
 
   const { title, description } = req.body;
 
@@ -38,22 +44,30 @@ const todos_post = async (req, res) => {
     res.status(201).json({ id: todo });
   } catch (error) {
     const errors = handleErrors(error);
-    res.status(400).send({ errors: errors });
+    res.status(500).send({ errors: errors });
     console.log(error);
   }
 };
 
 const update_post = async (req, res) => {
   const id = req.params.id;
-  const user = req.user
+  const user = req.user;
 
   const { title, description } = req.body;
 
   try {
-    const todo = await Todo.findByIdAndUpdate(id, { title, description, user });
-    res.status(201).json({ update: "Update successful" }); // Send JSON so frontend can handle redirect
+    if (id === req.user) {
+      const todo = await Todo.findByIdAndUpdate(id, {
+        title,
+        description,
+        user,
+      });
+      res.status(201).json({ update: "Update successful" });
+    } else {
+      res.status(500).send("This id is invalid");
+    }
   } catch (error) {
-    res.status(400).send("Error updating todo");
+    res.status(500).send("Error updating todo");
     console.log(error);
   }
 };
@@ -62,10 +76,14 @@ const delete_func = async (req, res) => {
   const id = req.params.id;
 
   try {
-    const todo = await Todo.findByIdAndDelete(id);
-    res.status(201).json({ delete: "Deleted todo successfully" });
+    if (id === req.user) {
+      const todo = await Todo.findByIdAndDelete(id);
+      res.status(201).json({ delete: "Deleted todo successfully" });
+    } else {
+      res.status(500).send("This id is invalid");
+    }
   } catch (error) {
-    res.status(400).send("Error deleting todo");
+    res.status(500).send("Error deleting todo");
     console.log(error);
   }
 };
